[transformer] Refacto transformer
