--- transformer/.coveragerc
@@ -1,2 +1,0 @@
-[run]
-omit = env/*,*__init__.py,tests/*
