--- transformer/Dockerfile
@@ -1,24 +1,0 @@
-FROM python:3.7-slim
-
-# Dependencies
-RUN apt-get update \
-    && apt-get -y install build-essential git
-
-ARG TRANSFORMER_HOME=/app
-ENV PYTHONPATH=${TRANSFORMER_HOME}
-
-WORKDIR ${TRANSFORMER_HOME}
-
-COPY transformer/requirements.txt ${TRANSFORMER_HOME}/requirements.txt
-
-RUN pip install -r ${TRANSFORMER_HOME}/requirements.txt
-
-COPY transformer/src ${TRANSFORMER_HOME}/transformer/src
-COPY analyzer/src ${TRANSFORMER_HOME}/analyzer/src
-COPY logger ${TRANSFORMER_HOME}/logger
-
-COPY transformer/uwsgi.ini ${TRANSFORMER_HOME}/transformer/uwsgi.ini
-
-COPY transformer/entrypoint.sh /entrypoint.sh
-
-CMD ["sh", "/entrypoint.sh"]
