--- transformer/entrypoint.sh
@@ -1,4 +1,0 @@
-#!/usr/bin/env bash
-
-# Run APP
-uwsgi --ini transformer/uwsgi.ini --py-autoreload 1
