--- transformer/requirements.txt
@@ -1,11 +1,0 @@
-arkhn-monitoring==0.0.4
-cleaning-scripts==0.2.27
-confluent_kafka==1.5.0
-fhir.resources @ git+https://github.com/arkhn/fhir.resources@6.0.1#egg=fhir.resources
-fluent-logger==0.9.6
-hiredis==1.1.0
-jsonschema==3.0.2
-prometheus-client==0.8.0
-redis==3.5.3
-requests==2.21.0
-uWSGI==2.0.18
