--- transformer/src/config/service_logger.py
@@ -1,3 +1,0 @@
-from logger import get_logger
-
-logger = get_logger(["resource_id", "primary_key_value"])
