--- transformer/unittest.cfg
@@ -1,13 +1,0 @@
-[unittest]
-plugins = nose2.plugins.junitxml
-
-[junit-xml]
-always-on = True
-
-[coverage]
-always-on = True
-coverage-report =
-	term 
-	xml
-	html
-coverage-config = .coveragerc
