--- transformer/uwsgi.ini
@@ -1,13 +1,0 @@
-[uwsgi]
-
-module = transformer.src.main:app
-master = true
-
-processes = 1
-threads = 4
-
-http = 0.0.0.0:3002
-
-vacuum = true
-
-die-on-term = true
